refactor: extract tag handling logic into TagsCommonFunctions and update imports
