+++ src/components/Post/Tags/components/index.tsx
@@ -1,5 +1,6 @@
 import { InputTag } from './_InputTag';
 import { useTagsLogic } from './_TagsUtils';
 import { InputTagStandard } from './_InputTagStandard';
+import * as TagsCommonFunctions from '../utils/_TagsCommonFunctions';
 
-export { InputTag, useTagsLogic, InputTagStandard };
+export { InputTag, useTagsLogic, InputTagStandard, TagsCommonFunctions };
